Fixed SpeakOnDemand not working
